Added support for the --js-client paramter
Added support for renderening all templates into one single javascript
file that can be included on the client.

it works by supplying the --js-client <namespace> parameter and then it
generates a single js file <namespace>.js that can be included, this
file contains a global object <namespace> that has a render function
which can be called to render a given template.